apps/server/pyproject.toml
[project]
name = "echo-bird-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "tortoise-orm[asyncpg]>=0.7.2",
    "aerich>=0.7.2",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.6",
    "bcrypt>=4.0.0",
    "twikit>=2.3.1",
    "cryptography>=41.0.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.395",
    "pytest>=8.4.0",
    "ruff>=0.11.13",
]


[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "PIE", # flake8-pie
    "SIM", # flake8-simplify
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",  # line too long (handled by line-length)
    "B008",  # do not perform function calls in argument defaults
    "RUF002",
    "RUF003"
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.aerich]
tortoise_orm = "app.database.TORTOISE_ORM"
location = "./app/migrations"
src_folder = "./"
